refactor(SignIn): extract logic to useSignInForm.ts
- Create useSignInForm custom hook to encapsulate form logic
- Move Zod schema and form input type definitions to useSignInForm.ts
- Implement useForm logic within the custom hook
- Add JSDoc comments for improved documentation
- Update SignIn component to use the new custom hook"

- Add Zod for schema validation
- Integrate zodResolver with react-hook-form for form validation
- Update SignIn component to use Zod schema for form validation
- Ensure proper error handling and display in form fields"
